Add PHPStan in CI (#1193)
* Add PHPStan

* Update php.yml

* Update composer.json

* Fix PHPStan level 0

* Fix PHPStan level 1

* Update phpstan.neon

* Fix PHPStan level 2

* Update composer.json

* Fix PHPStan level 3

* Fix tests

* Fix PHPStan level 4

* Update src/Common/Tests/TimedGeocoderTest.php


* Update src/Provider/Cache/Tests/ProviderCacheTest.php


* Update src/Provider/Cache/Tests/ProviderCacheTest.php


* Update src/Provider/GeoIP2/Tests/GeoIP2Test.php


* Fix PHPStan level 5

* Normalize composer.json

* Rename analyse script

* Update composer.json

* Update IntegrationTest

* Update AlgoliaPlacesTest

* Update composer.json

* Update RequestInterface vs. getParsedResponse()

* Update src/Plugin/PluginProvider.php


* Update src/Plugin/PluginProvider.php


* Use PHPStan baseline instead of ignore

See https://phpstan.org/user-guide/baseline

---------

+++ src/Plugin/Plugin/CachePlugin.php
@@ -79,8 +79,8 @@ class CachePlugin implements Plugin
     {
         if (null !== $this->precision && $query instanceof ReverseQuery) {
             $query = $query->withCoordinates(new Coordinates(
-                number_format($query->getCoordinates()->getLatitude(), $this->precision),
-                number_format($query->getCoordinates()->getLongitude(), $this->precision)
+                (float) number_format($query->getCoordinates()->getLatitude(), $this->precision),
+                (float) number_format($query->getCoordinates()->getLongitude(), $this->precision)
             ));
         }
